exercise: Udacity refresher on template classes
Templates are great when used with classes as it allows flexibility over the type of data that you can populate your class with. Note that if you want to specify arguments when creating a class instance, then we need to specify these variables in the private section, create a constructor function prototype, and finally define the contstructor function.